refactor(replay): rename sessionRecordingMetadataLogic -> sessionRecordingInfoLogic
The Metadata name overlapped with the existing sessionRecordingMetaLogic
in scenes/session-recordings/player/. Info sidesteps both that and the
backend's extra_summary_context overload.

Generated-By: PostHog Code
Task-Id: 10080d2e-9635-4f56-b7c1-8d9e69497395
